Change StockCount to int from uint

+++ CMP 205 Unit 1 WPF Application for Comic Book Store scenario/DundeeComicBookStore/Helpers/DBAccessHelper.cs
@@ -395,7 +395,7 @@ namespace DundeeComicBookStore
                         Name = (string)row["name"],
                         Description = (string)row["description"],
                         UnitPrice = (decimal)row["unitPrice"],
-                        UnitsInStock = (uint)(int)row["stockCount"]
+                        UnitsInStock = (int)row["stockCount"]
                     };
                     productList.Add(product);
                 }
@@ -437,7 +437,7 @@ namespace DundeeComicBookStore
                         Name = (string)row["name"],
                         Description = (string)row["description"],
                         UnitPrice = (decimal)row["unitPrice"],
-                        UnitsInStock = (uint)(int)row["stockCount"]
+                        UnitsInStock = (int)row["stockCount"]
                     };
                     productList.Add(product);
                 }
@@ -524,7 +524,7 @@ namespace DundeeComicBookStore
                     Name = (string)data["name"],
                     Description = (string)data["description"],
                     UnitPrice = (decimal)data["unitPrice"],
-                    UnitsInStock = (uint)(int)data["stockCount"],
+                    UnitsInStock = (int)data["stockCount"],
                     UnitCost = (decimal)data["unitCost"]
                 };
             }
